Spike for approach for cas3 controller -> service refactor
