+++ src/main/resources/static/cas3-api.yml
@@ -236,6 +236,34 @@ paths:
                 type: array
                 items:
                   $ref: '_shared.yml#/components/schemas/ReferralHistoryNote'
+  /premises/{premisesId}:
+    get:
+      tags:
+        - Premises
+      summary: Returns a premises
+      operationId: premisesPremisesIdGet
+      parameters:
+        - name: premisesId
+          in: path
+          description: ID of the premises to return
+          required: true
+          schema:
+            type: string
+            format: uuid
+      responses:
+        200:
+          description: successful operation
+          content:
+            'application/json':
+              schema:
+                $ref: 'cas3-schemas.yml#/components/schemas/Cas3Premises'
+        404:
+          description: invalid premises ID
+          content:
+            'application/json':
+              schema:
+                $ref: '_shared.yml#/components/schemas/Problem'
+
   /premises/summary:
     get:
       tags:
